+++ NEWS.md
@@ -1,3 +1,8 @@
+## rtables 0.5.1.6
+ * `paginate_table` now accepts `cpp` and will perform vertical pagination when it is non-null
+ * `vpaginate_table` is now deprecated
+ * Increased versioned dependency on `formatters` to `>=0.3.2.4`
+
 ## rtables 0.5.1.5
  * Support for section dividers (`section_div` argument in `split_rows_by*` function)
  * Updated versioned dependency on `formatters` to `>=0.3.2.3`